app/requirements.txt
python-dotenv==0.21.0
redis==4.4.0
flask[async]==2.2.2
